- patches.arch/x86-uv-fix-uv_hub_macro-bug.patch: x86 / UV:   Fix bug in uv_global_gru_mmr_address macro (bnc#579636,   FATE#306952).
suse-commit: c1a8e709a7a6edbc636fddeb663e5382fceb0b75

+++ arch/x86/include/asm/uv/uv_hub.h
@@ -339,7 +339,8 @@ static inline unsigned char uv_read_global_mmr8(int pnode, unsigned long offset)
  */
 static inline unsigned long uv_global_gru_mmr_address(int pnode, unsigned long offset)
 {
-	return UV_GLOBAL_GRU_MMR_BASE | offset | (pnode << uv_hub_info->m_val);
+	return UV_GLOBAL_GRU_MMR_BASE | offset |
+		((unsigned long)pnode << uv_hub_info->m_val);
 }
 
 /*
